% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verification_arguments.R
\name{CheckArguments}
\alias{CheckArguments}
\title{Vérification des arguments de la fonction principale du simulateur Natura}
\usage{
CheckArguments(
  file_arbre,
  file_etude,
  file_compile,
  horizon,
  dec_perturb,
  dec_tbe1,
  tbe1,
  dec_tbe2,
  tbe2,
  ht,
  vol
)
}
\arguments{
\item{file_arbre}{Nom du fichier contenant les informations sur les arbres et les placettes (table, csv ou xlsx). Le fichier doit contenir une ligne par arbre ou par classe de dhp/essence.
Si ce paramètre n'est pas utilisé, \code{file_compile} doit l'être. Le nom des colonnes doit être:
 \itemize{
   \item placette: identifiant unique de la placette
   \item sdom_bio: code du sous-domaine bioclimatique, en majuscule (ex: 2EST, 4OUEST ou 2E, 4O)
   \item altitude: altitude (m)
   \item type_eco: code du type écologique (ex: MS22)
   \item p_tot: précipitation totale annuelle, moyenne sur la période 1980-2010 (mm)
   \item t_ma: température moyenne annuelle sur la période 1980-2010 (Celcius)
   \item essence: code de l'essence de l'arbre (ex: SAB, EPN, BOP)
   \item dhpcm: dhp (cm) de l'arbre ou classe de dhp (seuls les arbres de plus de 9 cm sont retenus)
   \item tige_ha: nombre d'arbres de l'essence et de la classe de dhp, à l'hectare
   \item etat: code d'état de l'arbre, seuls les vivants sont retenus, codes 10, 12, 30, 32, 40, 42, 50, 52
   \item hauteur_pred: optionnel. hauteur totale de l'arbre (m). Si non fourni, la hauteur sera estimé à partir de la relation hauteur-diamètre de Auger (2016)
   \item vol_dm3: optionnel. volume marchand de l'arbre ou d'un arbre dans la classe de dhp (dm3). Si non fourni, le volume sera estimé à partir du tarif de cubage de Fortin et al. (2007)
 }}

\item{file_etude}{Nom du fichier contenant les arbres-études (table, csv ou xlsx) pour l'estimation de la hauteur dominante et l'âge du peuplement. Le fichier doit contenir une ligne par arbre-étude.
Si ce paramètre n'est pas utilisé, \code{file_compile} doit l'être. Le nom des colonnes doit être:
\itemize{
  \item placette: identifiant unique de la placette
  \item etage: code d'étage de l'arbre: C, D, O, I, V, seuls les C et D sont retenus
  \item essence: code de l'essence de l'arbre-étude (ex: SAB, EPN, BOP)
  \item dhpcm: dhp de l'arbre-étude (cm)
  \item hauteur: hauteur totale de l'arbre-étude (m)
  \item age: âge de l'arbre-étude, mesuré à 1 m (années)
}}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15}

\item{dec_perturb}{Numéro de la décennie avec une perturbation partielle (chablis partiel, brulis partiel), mettre 0 si aucune perturbation (0 par défaut)}

\item{dec_tbe1}{Numéro de la première décennie avec une épidémie de TBE, mettre 0 si aucune d'épidémie (0 par défaut)}

\item{tbe1}{Sévérité de la première épidémie, un entier de 1 à 5, mettre 0 si pas de TBE (0 par défaut)}

\item{dec_tbe2}{Numéro de la deuxième décennie avec une épidémie de TBE, mettre 0 si pas de 2e épidémie (0 par défaut)}

\item{tbe2}{Sévérité de la deuxième épidémie, un entier de 1 à 5, mettre 0 si pas de 2e épidémie (par défaut)}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): la hauteur de chacun des arbres doit être estimée
  \item \code{FALSE} : la variable hauteur_pred est fournie dans le fichier \code{file_arbre} (en m)
  }}

\item{vol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): le volume de chacun des arbres doit être estimé
  \item \code{FALSE}: la variable vol_dm3 est fournie dans le fichier \code{file_arbre} (en dm3 par tige), si le volume est fourni, la hauteur ne sera pas estimée.
}}
}
\value{
Une chaine de caractères contant "ok" s'il n'y a pas d'erreur, sinon, contient un message d'erreur.
}
\description{
Vérification des arguments de la fonction principale \code{SimulNatura()} du simulateur Natura-2014
}
