% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{espece}
\alias{espece}
\title{Table des essences traitées avec le regroupement}
\format{
## `espece`
Table de 44 lignes et 2 colonnes:
\describe{
  \item{essence}{Code d'essences}
  \item{groupe_ess}{Code du groupement d'essence}
  }
}
\usage{
espece
}
\description{
Table de 2 colonnes
}
\examples{
espece
}
\keyword{datasets}
