% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_arbres_etudes}
\alias{fichier_arbres_etudes}
\title{Fichier d'intrant des arbres-études pour Natura-2014}
\format{
## `fichier_arbres_etudes`
Table de 6 lignes et 6 colonnes:
\describe{
  \item{placette}{Identifiant de la placette}
  \item{ESSENCE}{Code de l'essence de l'arbre-étude}
  \item{etage}{Code d'étage de l'arbre-étude}
  \item{dhpcm}{DHP de l'arbre-étude (cm)}
  \item{hauteur}{Hauteur totale de l'arbre-étude (m)}
  \item{age}{Âge de l'arbre-étude, mesuré à 1 m (ans)}
}
}
\usage{
fichier_arbres_etudes
}
\description{
Deux placettes avec leur liste d'arbres-études
}
\examples{
fichier_arbres_etudes
}
\keyword{datasets}
