% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_arbres_sans_climat}
\alias{fichier_arbres_sans_climat}
\title{Fichier d'intrant à l'échelle de l'arbre pour Natura-2014, sans lew climat}
\format{
## `fichier_arbres_sans_climat`
Table de 29 lignes et 11 colonnes:
\describe{
  \item{placette}{Identifiant de la placette}
  \item{LATITUDE}{Latitude de la placette}
  \item{LONGITUDE}{Longitude de la placette}
  \item{essence}{Code de l'essence de l'arbre}
  \item{dhpcm}{Classe de DHP (cm)}
  \item{ETAT}{Code d'état de l'arbre}
  \item{TIGE_HA}{Nombre de tiges dans la classe de DHP et de l'essence, à l'hectare}
  \item{sdom_bio}{Code du sous-domaine bioclimatique}
  \item{REG_ECO}{Code de la région écologique}
  \item{type_eco}{Code du type écologique}
  \item{ALTITUDE}{Altitude (m)}
}
}
\usage{
fichier_arbres_sans_climat
}
\description{
Deux placettes avec leur liste d'arbres
}
\examples{
fichier_arbres_sans_climat
}
\keyword{datasets}
