% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_compile}
\alias{fichier_compile}
\title{Fichier d'intrant à l'échelle de la placette pour Natura-2014}
\format{
## `fichier_compile`
Table de 2 lignes et 25 colonnes:
\describe{
  \item{placette}{Identifiant de la placette}
  \item{sdom_bio}{Code du sous-domaine bioclimatique}
  \item{reg_eco}{Code de la région écologique}
  \item{type_eco}{Code du type écologique}
  \item{age}{Âge du peuplement (ans)}
  \item{is}{Indice de structure diamétrale de Shannon}
  \item{hd}{Hauteur dominante (m)}
  \item{ptot}{Précipitation totales annuelles (mm)}
  \item{t_ma}{Température annuelle moyenne  (C)}
  \item{altitude}{Altitude (m)}
  \item{nfi}{Densité marchande des feuillus intolérants (tiges/ha)}
  \item{nft}{Densité marchande des feuillus tolérants (tiges/ha)}
  \item{nrt}{Densité marchande des résineux tolérants (tiges/ha)}
  \item{nri}{Densité marchande des résineux intolérants (tiges/ha)}
  \item{nsab}{Densité marchande du sapin baumier (tiges/ha)}
  \item{stfi}{Surface terrière marchande des feuillus intolérants (m2/ha)}
  \item{stft}{Surface terrière marchande des feuillus tolérants (m2/ha)}
  \item{strt}{Surface terrière marchandedes résineux tolérants (m2/ha)}
  \item{stri}{Surface terrière marchandedes résineux intolérants (m2/ha)}
  \item{stsab}{Surface terrière marchande du sapin baumier (m2/ha)}
  \item{vfi}{Volume marchand des feuillus intolérants (m3/ha)}
  \item{vft}{Volume marchand des feuillus tolérants (m3/ha)}
  \item{vrt}{Volume marchand des résineux tolérants (m3/ha)}
  \item{vri}{Volume marchand des résineux intolérants (m3/ha)}
  \item{vsab}{Volume marchand du sapin baumier (m3/ha)}
}
}
\usage{
fichier_compile
}
\description{
Deux placettes
}
\examples{
fichier_compile
}
\keyword{datasets}
