% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphique.R
\name{generer_graphique}
\alias{generer_graphique}
\title{Graphique d'évolution des variables forestières}
\usage{
generer_graphique(Data, Espece = "tot", Variable = "st", listePlacette)
}
\arguments{
\item{Data}{Un data.frame contenant les données forestières. Le data.frame
doit inclure au moins les colonnes suivantes :
 placette (identifiant de la placette),
 annee,
 et les variables à tracer (par exemple, sttot, stbop, etc.).}

\item{Espece}{Une chaîne de caractères spécifiant l'espèce d'arbre pour
laquelle générer le graphique. Les valeurs possibles incluent
 "tot" (toutes essences),
 "bop" (bouleau à papier),
 "peu" (peupliers),
 "ft" (feuillus tolérants),
 "sab" (sapin baumier),
 "epn" (épinette noire),
 "epx" (autres épinettes),
 "ri" (résineux intolérants),
 "rt" (autres résineux tolérants).
 La valeur par défaut est "tot".}

\item{Variable}{Une chaîne de caractères spécifiant la variable à tracer.
Les valeurs possibles incluent
 "st" (surface terrière marchande en m2/ha),
 "n" (nombre d’arbres marchands par ha), "
 v" (volume marchand en m3/ha),
 "hd" (hauteur dominante en m),
 "dq" (diamètre quadratique moyen en cm).
 La valeur par défaut est "st".}

\item{listePlacette}{Une liste ou un vecteur contenant les identifiants des
placettes à inclure dans le graphique.}
}
\value{
Un objet ggplot représentant le graphique d'évolution de la variable
spécifiée sur la période d'année
}
\description{
Cette fonction génère un graphique d'évolution pour différentes
variables forestières sur une période d'année donnée, en fonction des
espèces et des placettes spécifiées.
}
