% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_compile.R
\name{Lecture_compile}
\alias{Lecture_compile}
\title{Lire le fichier des placettes à simuler avec Natura et valider le nom des colonnes}
\usage{
Lecture_compile(file_compile)
}
\arguments{
\item{file}{Nom du fichier des placette à lire (table ou Excel)}
}
\value{
Table à l'échelle de la placette ou un message d'erreur s'il y a une erreur dans le nom des colonnes.
}
\description{
Lire le fichier des placettes à simuler avec Natura et valider le nom des colonnes.
}
