% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_etudes.R
\name{Lecture_etudes}
\alias{Lecture_etudes}
\title{Lire le fichier des arbres-études et valider le nom des colonnes}
\usage{
Lecture_etudes(file_etude)
}
\arguments{
\item{file_etude}{Nom du fichier à lire (table ou Excel)}
}
\value{
Table dont les arbres-études ont été filtrés ou un message d'erreur s'il y a une erreur dans le nom des colonnes.
}
\description{
Lire le fichier des arbres-études et valider le nom des colonnes.
}
