% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Natura.R
\name{Natura}
\alias{Natura}
\title{Effectue une simulation complète (tous les pas de simulation) avec le modèle Natura-2014}
\usage{
Natura(
  data,
  horizon,
  dec_perturb = 0,
  dec_tbe1 = 0,
  tbe1 = 0,
  dec_tbe2 = 0,
  tbe2 = 0
)
}
\arguments{
\item{data}{Fichier compilé à la placette par groupe d'essences, départ des simulations}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15}

\item{dec_perturb}{Numéro de la décennie avec une perturbation partielle (chablis partiel, brulis partiel), mettre 0 si aucune perturbation (0 par défaut)}

\item{dec_tbe1}{Numéro de la première décennie avec une épidémie de TBE, mettre 0 si aucune d'épidémie (0 par défaut)}

\item{tbe1}{Sévérité de la première épidémie, un entier de 1 à 5, mettre 0 si pas de TBE (0 par défaut)}

\item{dec_tbe2}{Numéro de la deuxième décennie avec une épidémie de TBE, mettre 0 si pas de 2e épidémie (0 par défaut)}

\item{tbe2}{Sévérité de la deuxième épidémie, un entier de 1 à 5, mettre 0 si pas de 2e épidémie (par défaut)}
}
\value{
Table avec une ligne par placette et pas de simulation contenant les prévision de HD et IS et de N, ST et V de chacun des groupes d'essences
}
\description{
Effectue une simulation complète (tous les pas de simulation) avec le modèle Natura-2014
}
