% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Natura2014-package}
\alias{Natura2014}
\alias{Natura2014-package}
\title{Natura2014: Forest growth simulator Natura-2014}
\description{
Natura-2014 is a stand level forest growth model based on stand age. It uses merchantable stand characteristics at age t to predict them at time age+10 years. The model predict number of trees, basal area and volume for 5 species groups (tolerant hardwood, intolerant harwood, tolerant softwood, intolerant softwood, balsam fir). The model can only run in deterministic mode.
}
\author{
\strong{Maintainer}: Isabelle Auger \email{isabelle.auger@mrnf.gouv.qc.ca}

Other contributors:
\itemize{
  \item Mathieu Fortin \email{mathieu.fortin@nrcan-rncan.gc.ca} [contributor]
}

}
\keyword{internal}
