% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_arbres.R
\name{Prep_arbres}
\alias{Prep_arbres}
\title{Compilation des placettes à la step 0}
\usage{
Prep_arbres(fic_arbre, ht, vol)
}
\arguments{
\item{fic_arbre}{Table avec une ligne par arbre (ou par classe de dhp) pour chaque placette, avec les colonnes : placette, sdom_bio, altitude, p_tot, t_ma, essence, dhpcm, type_eco, tige_ha, etat}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): la hauteur de chacun des arbres doit être estimée
  \item \code{FALSE} : la variable hauteur_pred est fournie dans le fichier \code{file_arbre} (en m)
  }}

\item{vol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): le volume de chacun des arbres doit être estimé
  \item \code{FALSE}: la variable vol_dm3 est fournie dans le fichier \code{file_arbre} (en dm3 par tige), si le volume est fourni, la hauteur ne sera pas estimée.
}}
}
\value{
Table des caractéristiques des placettes (une ligne par placette)
}
\description{
Filtre les états, les essences et les dhp. Filtre les végétations potentielles et les sous-domaines bioclimatiques. Crée les variables veg_pet et milieu.
Estime la hauteur et le volume des arbres si nécessaire, calcule l'indice de Shannon et calcule les caractéristiques dendrométriques (N, ST, V) par groupe d'essences.
Calcule le dhp moyen des 4 plus gros arbres pour le calcul de la hauteur dominante.
}
