% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_etude.R
\name{Prep_etude}
\alias{Prep_etude}
\title{Calcule la hauteur dominante et l'âge moyen à la step 0}
\usage{
Prep_etude(fic_etude, compile_arbre)
}
\arguments{
\item{fic_etude}{Table des arbres-études}

\item{compile_arbre}{Table contenant les placettes compilées}
}
\value{
Table contenant la hauteur dominante et l'âge de chaque placette, une ligne par placette.
}
\description{
Calcule la hauteur dominante et l'âge moyen des placettes à la step 0 à partir des arbres-études, filtrer les arbres-études.
}
