% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_parametre.R
\name{Prep_parametre}
\alias{Prep_parametre}
\title{Lire les paramètres des équations de Natura-2014 et les joint au fichier des placettes}
\usage{
Prep_parametre(data)
}
\arguments{
\item{data}{Table contenant les placettes compilées}
}
\value{
Table contenant toutes les colonnes nécessaires pour effectuer une simulation, incluant les paramètres
}
\description{
Lit les paramètres des équations de Natura-2014 et les joint au fichier des placettes.
}
