% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_placettes.R
\name{valid_placette}
\alias{valid_placette}
\title{Vérifier le contenu des variables des fichier d'entrée}
\usage{
valid_placette(
  type_fic,
  fichier,
  ht = NULL,
  iqs = NULL,
  climat = NULL,
  sol = NULL
)
}
\arguments{
\item{type_fic}{Type du fichier à vérifier: arbres ou compil ou valid}

\item{fichier}{Nom de la table à vérifier}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): la hauteur de chacun des arbres doit être estimée
  \item \code{FALSE} : la variable hauteur_pred est fournie dans le fichier \code{file_arbre} (en m)
  }}

\item{climat}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables climatiques doivent être extraites des cartes. Les colonnes latitude et longitude de la placette doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables climatiques sont fournies dans le fichier \code{file_arbre} ou \code{file_compile}.
}}
}
\value{
Table ou message d'erreur
}
\description{
Vérifier le contenu des variables des fichiers d'entrée
}
