% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vp_retenues}
\alias{vp_retenues}
\title{Vecteur des végétations potentielles traitées}
\format{
## `vp_retenues`
}
\usage{
vp_retenues
}
\description{
Vecteur de 31
}
\examples{
vp_retenues
}
\keyword{datasets}
